Refactored directory structure
